+++ CHANGELOG.md
@@ -24,7 +24,9 @@
 
 * `track_alt`: allow **B2 Title** where _B2_ is followed by a space
 * `catalognum`: include **Catalog:** as a valid header
-* `track.title`: remove instances of `*digital-only`
+* `track.title` digital-only cleanup, remove:
+  * **DIGITAL** from the front of the title
+  * **digital-only** and **(digital)** from the end
 
 ### Fixed